make sure to also upload zip files
For windows we generate zip files, not tar.gz, so make sure to also
upload those.

Also upload the signature files a bit more explicitly.  Because we
have two sign processes running (first for building the release
artifacts and then the dev artifacts), we get .sig.sig files
otherwise, but we don't need these, so be more selective here.