Rewrite filestate.RenameStack logic in terms of checkpoints (#13927)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

The RenameStack function in filestate was one of the few places outside
of "cmd" still using `stack.DefaultSecretsProvider`, which it was only
using because the stack rename function was written in terms of
deserialised snapshots instead of serialised deployments.

This change rewrites the RenameStack function to work on the raw
deployment data rather than deserialised snapshot. This allows us to
remove the use of `stack.DefaultSecretsProvider` from filestate.

This would also allow the service to update their use of RenameStack to
work on the raw deployment as well, saving them the use of their
NoopSecretsProvider.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [ ] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->