Fix panic in --import-file with resources of the same resource name
Fixes #15002

Resources need to have unique names for the import file. To do this, a
map of taken names and the index they map to in the list of imported
resources is maintained.

While generating the import file, taken resource names and their
anticipated corresponding index would be added to the list before it was
appended. Between then and the append, the code could decide to
short-circuit leaving an invalid index behind.

The list of taken names is now only written to only if the resource is
added to the list of resource imports.