Temporarily revert validation of urn/id output properties (#15025)
Partial temporary revert of 8b263931631cd5f7d2ea66416c858ce3552946ed
(keeping a test that can be kept).

The validation is currently preventing the use of `id` as an output
property for packaged components (i.e. MLCs), which wasn't originally
considered when adding this validation. We do want to be able to support
`id` properties for components, so when we do subsequently add back the
validation, we'll do so in a way that doesn't prevent that.

Part of #15024