upgrade gocloud.dev to latest version (#15011)
Upgrade gocloud.dev to the latest version.  

This requires a bit of a workaround, since `gocloud.dev` changed its
expectations how the encryption key is stored for Azure. In v0.27.0 and
earlier, gocloud.dev accepted an encryption key that was wrapped in
base64.RawURLEncoding (and produced one that was wrapped as such.
However in v0.28.0 that changed and the encryption key was no longer
wrapped in gocloud.dev, and as such it also didn't expect a wrapped key
anymore .

To keep compatibility we'll keep wrapping the azure key for gocloud.dev
in an inner encoding, which seems to be the path of least resistance.
Alternatively we could introduce a `v2` encoding by prefixing the
string, but that ends up being messier than just keeping the strings
compatible.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
Tests were introduced in https://github.com/pulumi/pulumi/pull/14649,
I've updated them slightly to cover the new code as well.
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->