Temporary changes to unblock v3.100.0 release (#15050)
This commit contains two changes:

1. Temporarily comment out jobs that have already successfully run in
the `release.yml` workflow. See
https://github.com/pulumi/pulumi/actions/runs/7404894608

2. Add a way to manually dispatch a release via
`on-release-manual-dispatch.yml` workflow. The new file is a copy of
`on-release.yml`, but setup to allow manual trigger with input
variables, rather than being triggered by a release.

This change will be reverted after v3.100.0 is fully released.

Part of #15043