Revert "[sdk-gen/go] Fix compiling plain element type with plain maps" (#15059)
Reverts pulumi/pulumi#14704 

Fixes #15023 because the PR above changed how object types are
referenced when they are plain. It seems that we only support _shallow_
plain types (only container of nested objects can be plain, not the
object itself) which is why we confused the previous diffs to be
correct.

This reverts the PR previous to fix the breaking change at hand but will
reopen #14662 and make sure the fix for it is to generate `ObjectType`
_and_ `ObjectTypeArgs` regardless of whether the location where the
object type is referenced in the schema.