enable race detection in dev channel releases
We use dev channel releases to catch bugs early.  To increase the
chance bugs are found, also enable race detection in these releases.
This does mean these releases will be a little slower, and it might
introduce more errors, but that seems like a good thing in exchange
for potentially better quality of actual releases.