enable race detection for integeration tests
Enable race detection in the binary we're using for integration tests.
This will allow us to catch more data races before they get into any
release.  This does mean the binary we're using for integration tests
is slightly different from the binary we're releasing, however that's
already the case as we're running a binary with coverage enabled for
them.  Later we rebuild the binary we're actually releasing.