A property.Value implementation to replace resource.PropertyValue
The plan is to gradually replace resource.PropertyValue both internally (in
`pulumi/pulumi`) and in providers with this implementation.

This representation eliminates certain classes of bugs inherent with the old
representation:

- Modifiers (computed, secret, dependencies) live in a separate space from values, so
  there is no longer a problem with different nestings.

- Equality is well defined and encapsulated.

- The distinction between `Output`, `Secret` and `Computed`.

- Names have been normalized (Input/Computed, Map/Object).

This is our chance to have a property value representation that we like, so please leave
comments if you think we can improve the API here.