[python/sdk] Allow remote components to use output property called id (#15115)
# Description

This PR updates the python SDK to allow remote components to have ID
property. Adds an integration test which implements a component that had
an output property `id` to ensure that we can have `id` as an output
that doesn't get filtered out when serializing and deserializing
property maps.

Component instantiation looks like this in python
```python
import pulumi
from component import Component

component_a = Component("a", id="hello")

pulumi.export("id", component_a.id)
```
Where we expect `component_a.id` to equal `"{resource.ID}-{args.Id}" =>
"42-hello"` and the 42 is the (constant) ID of the resource created
inside the component.

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->