Move .version to /sdk and embed it
go:embed can not traverse paths upwards, move .version so we can embed
it. This enables us to have the version available outside of binary
builds (where we use linker flags to embed the version in
sdk/go/common/version/version.go).

This is useful for automation API, where we want to have the current
version available to us, but don't want to force consumers of the API to
build and test with linker flags.