[14573] Parse config paths strictly (#15173)
# Description

We want to parse config paths more strictly and not allow something like
"foo.[0]", but property paths everywhere else in the system need to
tolerate those keys.

Because of circular dependency on _pulumi/esc_ I kept a function with
old signature and added a new one. (instead of refactoring by adding new
parameter and exposing it)

Fixes https://github.com/pulumi/pulumi/issues/14573.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->

---------

Co-authored-by: Emil Hotkowski <ehotkowski@hopper.com>
Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>