[sdk/python] Install local SDK in automation API test
The test was installing a `requirements.txt` file that depends on the released version of `pulumi` package, but that package depends on the old `grpcio` package that does not work Python 3.12. Instead of depending on the public `pulumi` package, install the locally built Python SDK.

Also, fix the test to be able to run on Windows, while making changes here.