[sdk/python] Fix test to install local Pulumi SDK
The test is installing `pulumi-random` which currently depends on `pulumi`, which will fail on Python 3.12 because the currently released version of `pulumi` is pinned to a version of `grpcio` that doesn't work on Python 3.12.

Instead of depending on the public `pulumi` package, install the locally built Python SDK. Then install `pulumi-random`. That way, it'll be using a version of `grpcio` that works with Python 3.12.