Temporarily skip `TestConstructComponentConfigureProviderPython`
This test requires the `pulumi-tls` package, which currently depends on the published version of `pulumi` which can't be installed on Python 3.12 due to it requiring an older version of `grpcio` which can't be installed on Python 3.12.

The test does install the locally built Pulumi Python SDK, but `ProgramTest` does that _after_ installing any dependencies specified in the test program's `requirements.txt`.

We probably should consider updating `ProgramTest` to install the local dependencies first, or providing a new option to do that.

But in the meantime, the simplest workaround is to temporarily skip this test. After we've released a new version of `pulumi` that can be installed on Python 3.12, we can re-enable the test.